FHE precompile runtime access via FheApp struct (#11)
Instead of defining the runtime twice, we reference the one in the
FheApp struct. The idea is that it is less likely for the user to use a
runtime not associated with the FheApp instance.

Brings back to life #6 since we ended up being able to have Send + Sync
Applications.